fix layout
- Divide layout into public-facing and a layout for /admin
- Add proper ErrorBoundary
