+++ apps/frontend/package.json
@@ -30,11 +30,11 @@
     "@tailwindcss/postcss": "^4.1.10",
     "@tailwindcss/vite": "^4.0.3",
     "@tanstack/react-query": "^5.62.16",
-    "@tanstack/react-router": "^1.127.9",
-    "@tanstack/react-router-with-query": "^1.127.9",
-    "@tanstack/react-start": "^1.127.9",
+    "@tanstack/react-router": "^1.128.3",
+    "@tanstack/react-router-with-query": "^1.128.3",
+    "@tanstack/react-start": "^1.128.3",
     "@tanstack/react-table": "^8.20.6",
-    "@tanstack/router-plugin": "^1.127.9",
+    "@tanstack/router-plugin": "^1.128.3",
     "@uiw/react-codemirror": "^4.24.0",
     "@uiw/react-md-editor": "^4.0.7",
     "better-auth": "^1.2.12",
@@ -51,6 +51,7 @@
     "postcss": "^8.5.5",
     "react": "^19.0.8",
     "react-dom": "^19.0.3",
+    "react-error-boundary": "^6.0.0",
     "react-markdown": "^9.0.3",
     "rehype-raw": "^7.0.0",
     "rehype-slug": "^6.0.0",
@@ -65,9 +66,9 @@
     "@humandbs/eslint-config": "workspace:*",
     "@tailwindcss/typography": "^0.5.16",
     "@tanstack/eslint-plugin-query": "^5.66.0",
-    "@tanstack/eslint-plugin-router": "^1.114.12",
+    "@tanstack/eslint-plugin-router": "^1.125.0",
     "@tanstack/react-query-devtools": "^5.80.7",
-    "@tanstack/react-router-devtools": "^1.127.9",
+    "@tanstack/react-router-devtools": "^1.128.3",
     "@types/js-yaml": "^4.0.9",
     "@types/negotiator": "^0.6.4",
     "@types/pg": "^8.15.4",
