+++ apps/frontend/src/components/Skeleton.tsx
@@ -1,13 +1,12 @@
-export function Skeleton() {
+import { Skeleton } from "./ui/skeleton";
+
+export function SkeletonLoading() {
   return (
-    <div role="status" className="max-w-sm animate-pulse">
-      <div className="mb-4 h-2.5 w-48 rounded-full bg-slate-300"></div>
-      <div className="mb-2.5 h-2 max-w-[360px] rounded-full bg-slate-300"></div>
-      <div className="mb-2.5 h-2 rounded-full bg-slate-300"></div>
-      <div className="mb-2.5 h-2 max-w-[330px] rounded-full bg-slate-300"></div>
-      <div className="mb-2.5 h-2 max-w-[300px] rounded-full bg-slate-300"></div>
-      <div className="h-2 max-w-[360px] rounded-full bg-slate-300"></div>
-      <span className="sr-only">Loading...</span>
+    <div role="status" className="max-w-full min-w-sm animate-pulse space-y-2">
+      <Skeleton className="h-4 w-full" />
+      <Skeleton className="h-4 w-full" />
+      <Skeleton className="h-4 w-full" />
+      <Skeleton className="h-4 w-full" />
     </div>
   );
 }
