--- apps/frontend/src/routes/$lang/_layout/data-submission/application/index.tsx
@@ -1,11 +1,0 @@
-import { createFileRoute } from "@tanstack/react-router";
-
-export const Route = createFileRoute(
-  "/$lang/_layout/data-submission/application/"
-)({
-  component: RouteComponent,
-});
-
-function RouteComponent() {
-  return <div> Hello "/$lang/_layout/data-submission/application/"! </div>;
-}
