--- apps/frontend/src/routes/$lang/_layout/data-submission/navigation.somethng.tsx
@@ -1,16 +1,0 @@
-import { createFileRoute } from "@tanstack/react-router";
-
-export const Route = createFileRoute(
-  "/$lang/_layout/data-submission/navigation/somethng"
-)({
-  component: RouteComponent,
-  context({ context }) {
-    return {
-      crumb: "something",
-    };
-  },
-});
-
-function RouteComponent() {
-  return <div>Hello "/$lang/_layout/data-submission/navigation/somethng"!</div>;
-}
