--- apps/frontend/src/routes/$lang/_layout/guidelines/route.tsx
@@ -1,7 +1,0 @@
-import { createFileRoute } from "@tanstack/react-router";
-
-export const Route = createFileRoute("/$lang/_layout/guidelines")({
-  loader: () => ({
-    crumb: "guidelines",
-  }),
-});
