+++ apps/frontend/src/routes/_authed/admin.tsx
@@ -1,7 +1,4 @@
 import { Button } from "@/components/Button";
-import { Skeleton } from "@/components/Skeleton";
-import { Input } from "@/components/ui/input";
-import { Label } from "@/components/ui/label";
 import {
   Select,
   SelectContent,
@@ -9,27 +6,13 @@ import {
   SelectTrigger,
   SelectValue,
 } from "@/components/ui/select";
-import {
-  Table,
-  TableBody,
-  TableCell,
-  TableHead,
-  TableHeader,
-  TableRow,
-} from "@/components/ui/table";
+import { Skeleton } from "@/components/ui/skeleton";
 import { Tabs, TabsContent, TabsList, TabsTrigger } from "@/components/ui/tabs";
 import { ToggleGroup, ToggleGroupItem } from "@/components/ui/toggle-group";
 import { UserRole } from "@/db/schema";
 import { i18n, Locale } from "@/lib/i18n-config";
 import { roles } from "@/lib/permissions";
-import { useCopyToClipboard } from "@/lib/useColyToClipboard";
 import { cn } from "@/lib/utils";
-import { RenderMarkdoc } from "@/markdoc/RenderMarkdoc";
-import {
-  $deleteAsset,
-  $uploadAsset,
-  listAssetsQueryOptions,
-} from "@/serverFunctions/assets";
 import {
   $getDocuments,
   getDocumentsQueryOptions,
@@ -38,36 +21,17 @@ import {
   $createDocumentVersion,
   getDocumentVersionsListQueryOptions,
 } from "@/serverFunctions/documentVersion";
-import {
-  $createDocumentVersionTranslation,
-  $deleteDocumentVersionTranslation,
-  $updateDocumentVersionTranslation,
-  getDocumentVersionTranslationQueryOptions,
-} from "@/serverFunctions/documentVersionTranslation";
-import { config, processTokens, tokenizer } from "@/serverFunctions/getContent";
 import { $changeUserRole, getUsersQueryOptions } from "@/serverFunctions/user";
-import MDEditor from "@uiw/react-md-editor";
-import Markdoc from "@markdoc/markdoc";
 import {
   useMutation,
   useQueryClient,
   useSuspenseQuery,
 } from "@tanstack/react-query";
 import { createFileRoute } from "@tanstack/react-router";
-import {
-  ArrowLeft,
-  CopyIcon,
-  LucideTrash2,
-  PanelLeftClose,
-  PanelRight,
-  PanelRightClose,
-  PlusIcon,
-  Trash2Icon,
-  UploadIcon,
-} from "lucide-react";
-import { Suspense, useCallback, useEffect, useMemo, useState } from "react";
+import { Suspense, useState } from "react";
 import { useTranslations } from "use-intl";
-import { transformMarkdoc } from "@/markdoc/config";
+import { AssetsPanel } from "./-components/Assets";
+import { TranslationDetails } from "./-components/TranslationDetails";
 
 export const Route = createFileRoute("/_authed/admin")({
   component: RouteComponent,
@@ -76,17 +40,19 @@ export const Route = createFileRoute("/_authed/admin")({
 
 function RouteComponent() {
   return (
-    <Tabs defaultValue="news">
+    <Tabs defaultValue="news" className="z-50 flex-1">
       <TabsList>
         <TabsTrigger value="news">News</TabsTrigger>
         <TabsTrigger value="documents">Documents</TabsTrigger>
         <TabsTrigger value="users">Users</TabsTrigger>
       </TabsList>
-      <TabsContent value="news"></TabsContent>
-      <TabsContent value="documents">
-        <Suspense fallback={<div>Loading...</div>}>
-          <ManageDocuments />
-        </Suspense>
+      <TabsContent className="flex items-stretch gap-2" value="news">
+        <AssetsPanel />
+        <ManageNews />
+      </TabsContent>
+      <TabsContent className="flex items-stretch gap-2" value="documents">
+        <AssetsPanel />
+        <ManageDocuments />
       </TabsContent>
       <TabsContent value="users">
         <Suspense fallback={<div>Loading...</div>}>
@@ -97,12 +63,20 @@ function RouteComponent() {
   );
 }
 
-function ManageDocuments() {
-  const { data: documents } = useSuspenseQuery(getDocumentsQueryOptions());
-
-  const [selectedDocumentId, setSelectedDocumentId] = useState<string | null>(
-    null
+function ManageNews() {
+  return (
+    <Suspense fallback={<Skeleton />}>
+      <ListOfNews />
+    </Suspense>
   );
+}
+
+function ListOfNews() {
+  return <ul></ul>;
+}
+
+function ManageDocuments() {
+  const [selectedDocumentId, setSelectedDocumentId] = useState<string>();
 
   const [selectedLocale, setSelectedLocale] = useState<Locale>(
     i18n.defaultLocale
@@ -117,43 +91,44 @@ function ManageDocuments() {
     setSelectedVersionId(null);
   }
 
-  const t = useTranslations("Navbar");
-
   return (
-    <section className="flex items-stretch gap-2">
-      <AssetsPanel />
-      <ul className="bg-primary max-w-md space-y-4 rounded-sm p-4">
-        {documents.map((doc) => (
-          <li key={doc.id}>
-            <Button
-              className={cn({
-                "border-secondary-light border": doc.id === selectedDocumentId,
-              })}
-              onClick={() => handleSelectDoc(doc.id)}
-              variant={"toggle"}
-            >
-              {t(doc.contentId as any)}
-            </Button>
-          </li>
-        ))}
-      </ul>
+    <>
+      <div className="bg-primary w-md rounded-md p-4">
+        <Suspense fallback={<Skeleton />}>
+          <DocumentsList
+            onSelectDoc={handleSelectDoc}
+            selectedDocId={selectedDocumentId}
+          />
+        </Suspense>
+      </div>
 
       {selectedDocumentId ? (
         <>
-          <Suspense fallback={<div>Loading...</div>}>
-            <ListOfVersions
-              selectedVersionId={selectedVersionId}
-              onSelect={setSelectedVersionId}
-              documentId={selectedDocumentId}
-            />
-          </Suspense>
+          <div className="bg-primary rounded-sm p-2">
+            <Suspense
+              fallback={
+                <div>
+                  <Skeleton />
+                  <Button variant={"action"} disabled>
+                    Add new
+                  </Button>
+                </div>
+              }
+            >
+              <ListOfVersions
+                selectedVersionId={selectedVersionId}
+                onSelect={setSelectedVersionId}
+                documentId={selectedDocumentId}
+              />
+            </Suspense>
+          </div>
           {selectedVersionId ? (
             <div className="flex-1 rounded-sm bg-white p-4">
               <LocaleSwitcher
                 locale={selectedLocale}
                 onSwitchLocale={setSelectedLocale}
               />
-              <Suspense fallback={<div>Loading...</div>}>
+              <Suspense fallback={<Skeleton />}>
                 <TranslationDetails
                   locale={selectedLocale}
                   documentVersionId={selectedVersionId}
@@ -167,7 +142,55 @@ function ManageDocuments() {
       ) : (
         <div> No document selected </div>
       )}
-    </section>
+    </>
+  );
+}
+
+function DocumentsList({
+  onSelectDoc,
+  selectedDocId,
+}: {
+  onSelectDoc: (id: string) => void;
+  selectedDocId: string | undefined;
+}) {
+  const { data: documents } = useSuspenseQuery(getDocumentsQueryOptions());
+
+  const t = useTranslations("Navbar");
+
+  return (
+    <ul className="space-y-4">
+      {documents.map((doc) => (
+        <li key={doc.id}>
+          <Button
+            className={cn({
+              "border-secondary-light border": doc.id === selectedDocId,
+            })}
+            onClick={() => onSelectDoc(doc.id)}
+            variant={"toggle"}
+          >
+            {t(doc.contentId as any)}
+          </Button>
+        </li>
+      ))}
+    </ul>
+  );
+}
+
+function LocaleSwitcher({
+  locale,
+  onSwitchLocale,
+}: {
+  locale: Locale;
+  onSwitchLocale: (locale: Locale) => void;
+}) {
+  return (
+    <ToggleGroup type="single" value={locale} onValueChange={onSwitchLocale}>
+      {i18n.locales.map((loc) => (
+        <ToggleGroupItem key={loc} value={loc}>
+          {loc}
+        </ToggleGroupItem>
+      ))}
+    </ToggleGroup>
   );
 }
 
@@ -195,7 +218,7 @@ function ListOfVersions({
   }
 
   return (
-    <ul className="space-y-2 rounded-sm bg-white p-2">
+    <ul className="space-y-2">
       {versions.map((v) => (
         <li key={v.id}>
           <Button
@@ -215,127 +238,6 @@ function ListOfVersions({
         </Button>
       </li>
     </ul>
-  );
-}
-
-function LocaleSwitcher({
-  locale,
-  onSwitchLocale,
-}: {
-  locale: Locale;
-  onSwitchLocale: (locale: Locale) => void;
-}) {
-  return (
-    <ToggleGroup type="single" value={locale} onValueChange={onSwitchLocale}>
-      {i18n.locales.map((loc) => (
-        <ToggleGroupItem key={loc} value={loc}>
-          {loc}
-        </ToggleGroupItem>
-      ))}
-    </ToggleGroup>
-  );
-}
-
-function TranslationDetails({
-  documentVersionId,
-  locale,
-}: {
-  documentVersionId: string;
-  locale: Locale;
-}) {
-  const documentVersionTranslationQO =
-    getDocumentVersionTranslationQueryOptions({
-      documentVersionId,
-      locale,
-    });
-
-  const queryClient = useQueryClient();
-
-  const { data: versionDetails } = useSuspenseQuery(
-    documentVersionTranslationQO
-  );
-
-  const [value, setValue] = useState(versionDetails?.content);
-
-  useEffect(() => {
-    setValue(versionDetails?.content ?? "");
-  }, [versionDetails?.content]);
-
-  const { mutate: createTranslation } = useMutation({
-    mutationFn: $createDocumentVersionTranslation,
-    onSuccess: () => {
-      queryClient.invalidateQueries(documentVersionTranslationQO);
-    },
-  });
-
-  const { mutate: updateTranslation } = useMutation({
-    mutationFn: $updateDocumentVersionTranslation,
-    onSuccess: () => {
-      queryClient.invalidateQueries(documentVersionTranslationQO);
-    },
-  });
-
-  const { mutate: deleteTranslation } = useMutation({
-    mutationFn: $deleteDocumentVersionTranslation,
-    onSuccess: () => {
-      queryClient.invalidateQueries();
-    },
-  });
-
-  function handleSubmit() {
-    if (!value) return;
-    if (!versionDetails) {
-      createTranslation({
-        data: { documentVersionId, locale, content: value, title: "" },
-      });
-    } else {
-      updateTranslation({
-        data: {
-          documentVersionId,
-          locale,
-          content: value,
-        },
-      });
-    }
-  }
-
-  function handleDelete() {
-    deleteTranslation({
-      data: {
-        documentVersionId,
-        locale,
-      },
-    });
-  }
-
-  return (
-    <div className="flex flex-col gap-2">
-      <p> Author: {versionDetails?.translator?.name || "Unknown"} </p>
-      <div data-color-mode="light" className="flex-1">
-        <MDEditor
-          highlightEnable={true}
-          height="100%"
-          value={value}
-          onChange={setValue}
-          className="md-editor"
-          components={{
-            preview: (source) => {
-              const { content } = transformMarkdoc({ rawContent: source });
-
-              return <RenderMarkdoc content={content} />;
-            },
-          }}
-        />
-      </div>
-      <div className="flex h-fit justify-end gap-5">
-        <Button onClick={handleDelete} variant={"plain"} className="bg-white">
-          <LucideTrash2 className="text-red-600" />
-        </Button>
-        <Button onClick={handleSubmit} size={"lg"}>
-          {versionDetails ? "Update" : "Create"}
-        </Button>
-      </div>
-    </div>
   );
 }
 
@@ -376,182 +278,5 @@ function ManageUsers() {
         </li>
       ))}
     </ul>
-  );
-}
-
-function AssetsPanel() {
-  const [assetsOpen, setAssetsOpen] = useState(false);
-
-  return (
-    <div
-      className={cn(
-        "flex min-h-0 w-12 flex-col rounded-sm bg-white transition-all",
-        {
-          "w-assets-panel": assetsOpen,
-        }
-      )}
-    >
-      <Button
-        onClick={() => setAssetsOpen((prev) => !prev)}
-        className={cn("flex gap-2 text-black", { "flex-col": !assetsOpen })}
-        variant={"plain"}
-        size={"slim"}
-      >
-        {assetsOpen ? <PanelLeftClose /> : <PanelRight />}
-
-        <Label
-          className={cn("origin-left", {
-            "translate-x-5 -translate-y-2 rotate-90": !assetsOpen,
-          })}
-        >
-          Assets
-        </Label>
-      </Button>
-      {assetsOpen ? (
-        <Suspense fallback={<Skeleton />}>
-          <Assets />
-        </Suspense>
-      ) : null}
-    </div>
-  );
-}
-
-function Assets() {
-  const queryClient = useQueryClient();
-  const { data } = useSuspenseQuery(listAssetsQueryOptions());
-
-  const [filter, setFilter] = useState("");
-
-  const filteredAssets = useMemo(() => {
-    if (!data) return [];
-    if (!filter) return data;
-
-    return data.filter(
-      (asset) =>
-        asset.name.includes(filter) || asset.description.includes(filter)
-    );
-  }, [data, filter]);
-
-  const [addingNewAsset, setAddingNewAsset] = useState(false);
-  const [assetFile, setAssetFile] = useState<File>();
-  const [newAssetName, setNewAssetName] = useState("");
-
-  function handleChangeFile(e: React.ChangeEvent<HTMLInputElement>) {
-    const file = e.target.files?.[0];
-    if (!file) return;
-    setAssetFile(file);
-    setNewAssetName(file.name);
-  }
-
-  const { mutate } = useMutation({
-    mutationFn: $uploadAsset,
-    onSuccess: () => {
-      queryClient.invalidateQueries(listAssetsQueryOptions());
-    },
-  });
-
-  const { mutate: deleteAsset } = useMutation({
-    mutationFn: (id: string) => $deleteAsset({ data: { id } }),
-    onSuccess: () => {
-      queryClient.invalidateQueries(listAssetsQueryOptions());
-    },
-  });
-
-  function onSubmit(e: React.FormEvent<HTMLFormElement>) {
-    e.preventDefault();
-
-    const fd = new FormData(e.target as HTMLFormElement);
-
-    mutate({ data: fd });
-
-    setAssetFile(undefined);
-    setAddingNewAsset(false);
-  }
-
-  const [, copy] = useCopyToClipboard();
-
-  return (
-    <section className="flex flex-1 flex-col gap-2 p-4">
-      <Label>
-        <span className="text-sm font-normal whitespace-nowrap">
-          Search filter
-        </span>
-        <Input type="text" onChange={(e) => setFilter(e.target.value)} />
-      </Label>
-
-      <div className="min-h-0 flex-1 overflow-y-auto">
-        <Table className="max-h-full overflow-y-auto">
-          <TableHeader>
-            <TableRow>
-              <TableHead className="w-[100px]">Preview</TableHead>
-              <TableHead className="w-[100px]">Name</TableHead>
-              <TableHead>Type</TableHead>
-              <TableHead>Url</TableHead>
-              <TableHead className="text-right">Action</TableHead>
-            </TableRow>
-          </TableHeader>
-          <TableBody>
-            {filteredAssets.map((row) => (
-              <TableRow key={row.id}>
-                <TableCell>
-                  <img src={row.url} />
-                </TableCell>
-                <TableCell>{row.name}</TableCell>
-                <TableCell>{row.mimeType}</TableCell>
-                <TableCell>
-                  <Button
-                    variant={"cms-table-action"}
-                    className="text-black"
-                    onClick={() => copy(row.url)}
-                  >
-                    <CopyIcon className="size-4" />
-                  </Button>
-                </TableCell>
-                <TableCell className="text-right">
-                  <Button
-                    onClick={() => deleteAsset(row.id)}
-                    variant={"cms-table-action"}
-                  >
-                    <Trash2Icon className="text-accent size-4" />
-                  </Button>
-                </TableCell>
-              </TableRow>
-            ))}
-          </TableBody>
-        </Table>
-      </div>
-      <Button
-        disabled={addingNewAsset}
-        variant={"toggle"}
-        onClick={() => setAddingNewAsset(true)}
-      >
-        <PlusIcon className="mr-2 inline size-4" /> Add new asset
-      </Button>
-      {addingNewAsset ? (
-        <form onSubmit={onSubmit} className="space-y-1">
-          <Input
-            name="file"
-            type="file"
-            accept="image/*,application/pdf"
-            onChange={handleChangeFile}
-          />
-          <Input
-            name="name"
-            placeholder="Name"
-            value={newAssetName}
-            onChange={(e) => setNewAssetName(e.target.value)}
-          />
-          <Input name="description" placeholder="Description" />
-          <Button
-            disabled={!assetFile || !newAssetName}
-            type="submit"
-            className="ml-auto"
-          >
-            <UploadIcon className="mr-2 inline size-4" />
-            Upload
-          </Button>
-        </form>
-      ) : null}
-    </section>
   );
 }
