+++ apps/frontend/src/routes/_main/$lang/_layout/achievements/index.tsx
@@ -1,0 +1,10 @@
+import { createFileRoute } from "@tanstack/react-router"
+
+export const Route = createFileRoute("/_main/$lang/_layout/achievements/")({
+  component: RouteComponent,
+  pendingComponent: () => <div>Loading...</div>,
+})
+
+function RouteComponent() {
+  return <div>Hello "/results/"!</div>
+}
