+++ apps/frontend/src/routes/_main/$lang/_layout/data-submission/navigation.somethng.tsx
@@ -1,0 +1,16 @@
+import { createFileRoute } from "@tanstack/react-router";
+
+export const Route = createFileRoute(
+  "/_main/$lang/_layout/data-submission/navigation/somethng"
+)({
+  component: RouteComponent,
+  context({ context }) {
+    return {
+      crumb: "something",
+    };
+  },
+});
+
+function RouteComponent() {
+  return <div>Hello "/$lang/_layout/data-submission/navigation/somethng"!</div>;
+}
