+++ apps/frontend/src/routes/_main/$lang/_layout/guidelines/route.tsx
@@ -1,0 +1,7 @@
+import { createFileRoute } from "@tanstack/react-router";
+
+export const Route = createFileRoute("/_main/$lang/_layout/guidelines")({
+  loader: () => ({
+    crumb: "guidelines",
+  }),
+});
